Both busData and branchData now work.
But the cdf.txt info seems to be mostly inconsistent across bus systems. MATPOWER's cdf2mpc just ignores the last six (out of a total of 21) columns of the branchData, possibly because of inconsistency in column widths (and most probably because the data in them is dispensible). My branchData will work for a reduced fifteen columns with the reducedBranchColumns option. A verbose option is also added to the function.