Important change in computation of PSpec, QSpec made.
In initializeVectors, I was using P_G(i) for PV buses an -P_D(i) for PQ buses as the values for PSpec, and similarly for QSpec. But It should be P_G(i)-P_D(i) for all buses. Apart from that, the Kothari 3 bus system has been added.