Supressed warnings in constructSparseJacobian
Extension of a table by specifying one element at a time is NOT liked by MATLAB. It seems to want us to define the whole next row in one command. But for me upon trying to define something like table{newlyAddedRowNum, :} = [nrow, ncol, val] it always demotes my val(s) from 'double' to 'int64', which is NOT acceptable.